requirements.txt
numpy==1.24.3
Pillow==9.4.0
torch==2.0.1
torchvision==0.15.2
